Avoiding compiler warnings on Visual Studio.
We use some old C functions like 'strcat' and 'fopen' that windows
consider insecure (like other Operating Systems). But Windows
replacement for some of these functions are not very portable. We
should aim to replace 'strcpy' and 'strcat' for 'memcpy' because it
appears to be the most portable replacement. But anyway we need to
suppress these warnings in Windows because we do not plan to replace
'fopen'.